clean up LGP test config
useSwitches does not exist anymore since
8c8337214f7a33c566409308a159e0930819b891 .

Hrate and FlagConstraints only exist in retired code or comments.

TM_Physics not at all.

collisions is a boolean, and enabled by default.
Keeping it commented out looks like it might be disabled by default..

"problem" exists in some tools and python bindings, but not in the C++
codebase.